#5693: move context to better location
